requirements.txt
asgiref==3.11.0
Django==5.2.8
djangorestframework==3.15.2
Pillow==11.0.0
sqlparse==0.5.3
